pyproject.toml
[project]
name = "forensic-psych-report-drafter"
version = "0.1.0"
description = "An experimental forensic report drafter"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic[bedrock]==0.84.0",
    "azure-ai-documentintelligence==1.0.2",
    "google-genai==1.66.0",
    "genai-prices==0.0.55",
    "langchain-text-splitters==1.1.1",
    "openai==2.26.0",
    "psutil==7.2.2",
    "pymupdf==1.27.1",
    "python-dotenv==1.2.2",
    "python-frontmatter==1.1.0",
    "tenacity==9.1.4",
    "tiktoken==0.12.0",
    "python-docx==1.2.0",
    "keyring==25.7.0",
    "pypandoc==1.16.2",
    "pypandoc-binary==1.16.2",
    "pyside6==6.10.2",
    "pydantic-ai-slim==1.67.0",
    "docling[vlm]>=2.80.0",
    "mlx>=0.31.1",
    "mlx-lm>=0.29.1",
]

[dependency-groups]
dev = [
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "pytest-qt==4.5.0",
    "arize-phoenix==13.9.0",
    "openinference-instrumentation==0.1.46",
    "openinference-instrumentation-openai==0.1.41",
    "openinference-instrumentation-anthropic==1.0.0",
    "rich==14.3.3",
    "pyinstaller==6.19.0",
]

[tool.pytest]
addopts = ["-ra", "--strict-markers", "--strict-config"]
strict = true
qt_api = "pyside6"
testpaths = ["tests"]
markers = [
    "unit: fast deterministic unit tests",
    "core: dashboard core/domain tests",
    "worker: worker and orchestration tests",
    "ui: Qt UI and controller/widget tests",
    "integration: multi-module integration tests",
    "live_provider: tests that call real external LLM providers",
    "slow: tests that are intentionally long-running",
]
filterwarnings = [
    "ignore::DeprecationWarning:phoenix.server.api.routers",
]
